+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
   - https://github.com/Shopify/flash-list/pull/412
 - Define `FlashList` props previously inherited from `VirtualizedList` and `FlatList` explicitly
   - https://github.com/Shopify/flash-list/pull/386
+- Make `estimatedItemSize` optional
+  - https://github.com/Shopify/flash-list/pull/378
 - Change `overrideItemType` prop name to `getItemType`
   - https://github.com/Shopify/flash-list/pull/369
 - Added `useBlankAreaTracker` hook for tracking blank area in production